refactor(cli): postObj to create
